Implement Algorithm T
This is a rendering of the algorithm on page 265 into TypeScript.
The steps of the algorithm appear as comments in the source code.
Some care has been taken to stay close to the names from the
algorithm.

Although not strictly necessary, we also do some bounds checking
on the relations in the input. In the algorithm, those bounds are
assumed to hold.