Improve makeCircularList function
The old version is to be considered suspect and buggy.

For one thing, I realized the book wants the `ptr` to point to the
_rightmost_ ("end") node in the circular list. The way I had it in
the previous version, it pointed to the _leftmost_ end. Facepalm.

But then I also found a way to express things much shorter and clearer.
I got rid of the type `ProtoCircularList`, which simply isn't needed.
I also got rid of the special case of an empty circular list, which
is now simply handled on-the-fly by checking if we even _have_ a
`rightmostNode`. Along the way, we also got more suitable variable
names.

I also discovered I had made a horrible mistake inside the
`reduceRight`, causing the whole `makeCircularList` function not
to work up until this point. Short story: I was trying to do two things
at the same time: (a) setting `rightmostNode` (formerly known as `last`)
and (b) returning a newly created node. Somehow I screwed up the latter.

Yep, I will hand in my "Master" badge and magical Staff of Wisdom at
the front desk, and then walk on foot back home across the desert.